Code/Operationalization/pjme/requirements.txt
asgiref==3.2.10
attrs==19.3.0
autobahn==20.7.1
Automat==20.2.0
autopep8==1.5.4
backcall==0.2.0
bleach==3.1.5
boto3==1.14.38
botocore==1.17.38
Brotli==1.0.7
certifi==2020.6.20
cffi==1.14.2
channels==2.4.0
chardet==3.0.4
click==7.1.2
cmdstanpy==0.4.0
colorama==0.4.3
constantly==15.1.0
convertdate==2.2.1
cryptography==3.1
cycler==0.10.0
cypari==2.4.0
Cython==0.29.21
daphne==2.5.0
dash==1.14.0
dash-core-components==1.10.2
dash-html-components==1.0.3
dash-renderer==1.6.0
dash-table==4.9.0
decorator==4.4.2
defusedxml==0.6.0
dj-database-url==0.4.1
Django==3.1
django-filter==2.3.0
django-pandas==0.6.2
django-plotly-dash==1.4.2
django-secrets==1.0.2
django-secrets-manager==0.1.13
djangorestframework==3.11.0
docutils==0.15.2
dpd-components==0.1.0
entrypoints==0.3
ephem==3.7.7.1
fastparquet==0.4.1
fbprophet==0.6
Flask==1.1.2
Flask-Compress==1.5.0
freeze==3.0
future==0.18.2
FXrays==1.3.3
gunicorn==19.6.0
holidays==0.10.3
hyperlink==20.0.1
idna==2.10
importlib-metadata==1.7.0
incremental==17.5.0
ipykernel==5.3.4
ipython==7.17.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
jsonschema==3.2.0
jupyter-client==6.1.6
jupyter-core==4.6.3
kiwisolver==1.2.0
korean-lunar-calendar==0.2.1
llvmlite==0.33.0
LunarCalendar==0.0.9
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.0
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.4
nltk==3.2.1
numba==0.50.1
numpy==1.19.1
oauthlib==1.1.2
packaging==20.4
pandas==1.1.0
pandocfilters==1.4.2
parso==0.7.1
pickleshare==0.7.5
Pillow==7.2.0
plink==2.3.1
plotly==4.9.0
prompt-toolkit==3.0.5
psycopg2==2.7.5
pyarrow==1.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.6.1
PyHamcrest==2.0.2
PyMeeus==0.3.7
pyOpenSSL==19.1.0
pyparsing==2.4.7
pypng==0.0.20
pyreadline==2.1
pyrsistent==0.16.0
pystan==2.19.1.1
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
pywin32==228
pyzmq==19.0.2
requests==2.11.1
requests-oauthlib==0.6.2
retrying==1.3.3
s3transfer==0.3.3
scikit-learn==0.23.2
scipy==1.5.2
seaborn==0.10.1
service-identity==18.1.0
setuptools-git==1.2
six==1.10.0
sklearn==0.0
snappy-manifolds==1.1.1
spherogram==1.8.3
sqlparse==0.3.1
testpath==0.4.4
textblob==0.11.1
threadpoolctl==2.1.0
thrift==0.13.0
toml==0.10.1
tornado==6.0.4
traitlets==4.3.3
tweepy==3.5.0
Twisted==20.3.0
txaio==20.4.1
Unipath==1.1
urllib3==1.25.10
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==5.0
zipp==3.1.0
zope.interface==5.1.0